massively refactored event-resources.jsx
