+++ pages/event-resources.jsx
@@ -1,415 +1,2 @@
-var React = require('react');
-var ImageTag = require('../components/imagetag.jsx');
-var Router = require('react-router');
-var Link = Router.Link;
-
-var config = require('../config/config');
-
-var PageLinker = React.createClass({
-  render: function() {
-    return (
-      <li className="page-linker">
-        <a href={this.props.href}>{this.props.head}</a>
-      </li>
-    );
-  }
-});
-var EventItem = React.createClass({
-  propTypes: {
-    participants: React.PropTypes.string.isRequired,
-    linkToGuide: React.PropTypes.string.isRequired,
-    head: React.PropTypes.string.isRequired,
-    subHead: React.PropTypes.string.isRequired,
-    content: React.PropTypes.string.isRequired
-  },
-  render: function() {
-    return (
-      <div className="event-item">
-        <div className="event-icon">
-          <div className="participants-number">{this.props.participants}</div>
-          <div className="participants-label">participants</div>
-        </div>
-        <div className="event-content">
-          <div className="event-item-header"><a href={this.props.linkToGuide}>{this.props.head}</a></div>
-          <div className="event-item-subheader">{this.props.subHead}</div>
-          <p className="event-item-content">{this.props.content}</p>
-        </div>
-      </div>
-    );
-  }
-});
-var LogoAsset = React.createClass({
-  getInitialState: function() {
-    return { hasKeyboardFocus: false };
-  },
-  handleFocus: function() {
-    this.setState({ hasKeyboardFocus: true });
-  },
-  handleBlur: function() {
-    this.setState({ hasKeyboardFocus: false });
-  },
-  render: function() {
-    return (
-      <div className="logo-asset-container col-sm-4 col-md-4 col-lg-3">
-        <div className="logo-asset-header">{this.props.head}</div>
-        <ImageTag className="logo-asset-img"
-        width={200} height={200}
-        src1x={this.props.src1x} src2x={this.props.src2x}
-        alt={this.props.alt}/>
-        <div className={"logo-asset-hover " + (
-          this.state.hasKeyboardFocus ? "has-keyboard-focus" : ""
-         )} onFocus={this.handleFocus} onBlur={this.handleBlur}>
-          <div className="logo-asset-center">{this.props.children}</div>
-        </div>
-      </div>
-    );
-  }
-});
-var LogoAssetLink = React.createClass({
-  render: function() {
-    return (
-      <a className="logo-asset-link" href={this.props.href}><span className="fa fa-download">&nbsp;</span>{this.props.children}</a>
-    );
-  }
-});
-var RemixLink = React.createClass({
-  render: function() {
-    return (
-      <a className="remix-link" href={this.props.href}>Remix</a>
-    );
-  }
-});
-var Tabulator = React.createClass({
-  getInitialState: function() {
-    return {
-      activeClass: "tab-0"
-    };
-  },
-  propTypes: {
-    tabs: React.PropTypes.array
-  },
-  showTab: function(activeTab) {
-    this.setState({
-      activeClass: activeTab
-    });
-  },
-  render: function() {
-    var tabulator = this;
-    var className = "tabulator";
-    if (this.state.activeClass) {
-      className += " " + this.state.activeClass;
-    }
-    return (
-      <div className={className}>
-        <div className="tabulator-head-container">
-          {this.props.tabs.map(function (section, key) {
-            function onClick(e) {
-              tabulator.showTab("tab-" + key);
-            }
-            var className = "tabulator-head col-xs-3 col-sm-3 col-md-3 col-lg-3";
-            className += " tab-" + key;
-            return (
-              <div onClick={onClick} className={className} key={key}>{section.head}</div>
-            );
-          })}
-        </div>
-        {this.props.tabs.map(function (section, key) {
-          var className = "tabulator-content-container";
-          className += " tab-" + key;
-          return (
-            <div key={key}>
-              <div className="tabulator-head-no-js">{section.head}</div>
-              <div className={className}>
-                <div className="tabulator-content">
-                  {section.content}
-                </div>
-              </div>
-            </div>
-          );
-        })}
-      </div>
-    );
-  }
-});
-var EventDetail = React.createClass({
-  render: function() {
-    return (
-      <div>
-        <div className="event-detail-head col-sm-3 col-md-3 col-lg-3">{this.props.head}</div>
-        <div className="event-detail-content col-sm-9 col-md-9 col-lg-9">{this.props.children}</div>
-      </div>
-    );
-  }
-});
-
-var EventsResources = React.createClass({
-  statics: {
-    pageTitle: 'Event Resources',
-    pageClassName: 'event-resources',
-    LogoAsset: LogoAsset
-  },
-  render: function() {
-    return (
-      <div className="inner-container">
-        <h1>Event Resources</h1>
-        <p>Thank you for helping us celebrate webmaking around the world! These resources will help you plan a unique event tailored especially for your audience.</p>
-
-        <div className="row">
-          <ul className="caret-list">
-            <PageLinker head="event guides" href="#event-guides"></PageLinker>
-            <PageLinker head="logos & assets" href="#logo-assets"></PageLinker>
-            <PageLinker head="event details" href="#event-details"></PageLinker>
-            <PageLinker head="event support" href="#event-support"></PageLinker>
-          </ul>
-        </div>
-
-        <h2 id="event-guides">Event Guides</h2>
-        <p>No matter the size of your event, we have a guide for you.</p>
-
-        <EventItem participants="2-5"
-          linkToGuide="https://michelle.makes.org/thimble/LTE2NjE1MzQyMDg=/how-to-host-a-maker-party-small"
-          head="small event"
-          subHead="Perfect for 2 to 5 participants"
-          content="A fun way to spend an hour on a rainy day, hang out as a family, learn to hack with a friend, and make cool things on the web."
-        />
-        <EventItem participants="5-50"
-          linkToGuide="https://michelle.makes.org/thimble/LTE1Nzc2NDgxMjg=/how-to-host-a-maker-party-medium"
-          head="medium event"
-          subHead="Great for 5 to 50 participants"
-          content="A fantastic way to team up people with different skill-sets to collaboratively build something new or improve something existing on the web, all while learning and teaching new skills."
-        />
-        <EventItem participants="50+"
-          linkToGuide="https://michelle.makes.org/thimble/LTEzMDkyMTI2NzI=/how-to-host-a-maker-party-large"
-          head="large event"
-          subHead="For those ready for the adventure of 50+ participants"
-          content="A rewarding way to bring together local organizations in a science fair setting to demonstrate cool web ideas, provide fun hands-on activities, and introduce your community to making and hacking."
-        />
-
-        <h2 id="logo-assets">Logos & Assets</h2>
-        <p>Feel free to use these Maker Party graphics in any of you promotional materials:</p>
-
-        <div className="row">
-          <LogoAsset head="Maker Party Logo"
-          alt="Maker Party Logo Image"
-          src1x="/img/pages/event-resources/resource-thumbnails-01.png"
-          src2x="/img/pages/event-resources/resource-thumbnails-01@2x.png">
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyLogo.eps">Download EPS</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyLogo.png">Download PNG</LogoAssetLink>
-          </LogoAsset>
-          <LogoAsset head="Brand Palette"
-          alt="Brand Palette Image"
-          src1x="/img/pages/event-resources/resource-thumbnails-03.png"
-          src2x="/img/pages/event-resources/resource-thumbnails-03@2x.png">
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyBrandPalette.eps">Download EPS</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyBrandPalette.png">Download PNG</LogoAssetLink>
-          </LogoAsset>
-          <LogoAsset head="Poster"
-          alt="Poster Image"
-          src1x="/img/pages/event-resources/resource-thumbnails-07.png"
-          src2x="/img/pages/event-resources/resource-thumbnails-07@2x.png">
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyBanner.eps">Download EPS</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyBanner.png">Download PNG</LogoAssetLink>
-          </LogoAsset>
-          <LogoAsset head="Desktop Wallpaper"
-          alt="Desktop Wallpaper Image"
-          src1x="/img/pages/event-resources/resource-thumbnails-10.png"
-          src2x="/img/pages/event-resources/resource-thumbnails-10@2x.png">
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-320x480.jpg">320px x 480px</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-640x1136.jpg">640px x 1136px</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-768x1280.jpg">768px x 1280px</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-1920x1200.jpg">1920px x 1200px</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-2560x1440.jpg">2560px x 1440px</LogoAssetLink>
-          </LogoAsset>
-          <LogoAsset head="Certificate"
-          alt="Certificate Image"
-          src1x="/img/pages/event-resources/resource-thumbnails-08.png"
-          src2x="/img/pages/event-resources/resource-thumbnails-08@2x.png">
-            <RemixLink href="https://d157rqmxrxj6ey.cloudfront.net/mhasan/13029/ ">Download EPS</RemixLink>
-          </LogoAsset>
-          <LogoAsset head="Table Cloth Design"
-          alt="Table Cloth Design Image"
-          src1x="/img/pages/event-resources/resource-thumbnails-04.png"
-          src2x="/img/pages/event-resources/resource-thumbnails-04@2x.png">
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyTableClothDesign.eps">Download EPS</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyTableClothDesign.png">Download PNG</LogoAssetLink>
-          </LogoAsset>
-          <LogoAsset head="Buttons"
-          alt="Buttons Image"
-          src1x="/img/pages/event-resources/resource-thumbnails-06.png"
-          src2x="/img/pages/event-resources/resource-thumbnails-06@2x.png">
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyButtons.eps">Download EPS</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyButtons.png">Download PNG</LogoAssetLink>
-          </LogoAsset>
-          <LogoAsset head="KUMI Papertoy"
-          alt="KUMI Papertoy Image"
-          src1x="/img/pages/event-resources/resource-thumbnails-09.png"
-          src2x="/img/pages/event-resources/resource-thumbnails-09@2x.png">
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/KUMI-Papertoy-yellow.pdf">Yellow T-sirt</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/KUMI-Papertoy-blue.pdf">Teal T-shirt</LogoAssetLink>
-            <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/KUMI-Papertoy-blank-tees.pdf">Design Your Own T-shirt</LogoAssetLink>
-          </LogoAsset>
-        </div>
-
-        <h2 id="event-details">Event Details</h2>
-        <p>There are many details to consider when planning your event. Here's a list of what you need to know:</p>
-
-        <Tabulator tabs={[
-          {
-            "head": "the basics",
-            "content": (
-              <div>
-                <EventDetail head="Find a Venue">
-                  <div>
-                    Look for a comfortable, flexible space that's appropriate for the number of participants you expect. Contact local community spaces, museums, hackerspaces, coworking spaces, organizations, libraries, schools, or coffee shops to see if that have space available for events. Send them information on your event, the missions and what you hope to achieve and offer opportunities for their community to attend. Ensure there is reliable internet, enough power outlests and good lighting.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Prepare Equipment">
-                  <div>
-                    Create a list of what equipment you need for your event. Check out the event space ahead of time to learn what additional items you might need to buy or bring. If learners need to supply their own equipment, like laptops, make sure this is communicated clearly ahead of time. Also reach out to your venue, mentors or other local organizations to see if they can lend equipment.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Gather Materials">
-                  <div>
-                    Since we know there will be note-taking, brainstorming and making, you are going to need some materials. Have a sign-up sheet, name tags, post-it notes, pens and paper easily accessible. Prepare any other materials you will need ahead of time and set a table aside at the event where learners can grab materials as needed.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Make Some Gear">
-                  <div>
-                    Use this <a href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyLogo.eps">logo</a> to create Maker Party t-shirts at your local print shop. Here's an example of what your <a href="https://stuff.webmaker.org/teach.mozilla.org/makerparty_tshirtpreview.jpg">shirts</a> could look like. Don't have all the budget for printing? Ask participants to bring their own shirts and create Maker Party t-shirts with <a href="https://tbx.makes.org/thimble/stencil-a-tshirt">this activity</a>. You can also print these <a href="https://makerparty2014.makes.org/thimble/LTEzNDM3NTAxNDQ=/maker-party-love-bomb">certificates</a> to give to the participants at the event that successfully complete the activities.
-                  </div>
-                </EventDetail>
-                <EventDetail head="All About The Wi-Fi">
-                  <div>
-                    Have you ever heard of an event with perfect Wi-Fi? Neither have we. Prepare for the worst and have a back-up plan in case you lose connection. There are also a lot of <a href="https://keyboardkat.makes.org/thimble/LTIxMDA3NTY0ODA=/lofi-nofi-teaching-kit">activities that require no Internet access</a>. Print out a few activities and keep them with you just in case.
-                  </div>
-                </EventDetail>
-              </div>
-            )
-          },
-          {
-            "head": "before event",
-            "content": (
-              <div>
-                <EventDetail head="Recruit Mentors">
-                  <div>
-                    Whether they’re colleagues, students or a group of talented friends, you can find volunteer mentors by tapping into your own networks. Provide them a <a href="https://stuff.webmaker.org/teach.mozilla.org/Sample_Volunteer_Package.pdf">Sample Volunteer Package</a> with information on the event, activities, social media and any other necessary information in advance. Depending on how many volunteers you have, you might consider holding a meet-up in person or online prior to the event to answer any questions. Don't forget to gather and distribute contact information so that you can stay in touch with them in the days and weeks leading up to the event.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Promote Your Event">
-                  <div>
-                    Start spreading the word - share the link to your event so people can plan to attend or share it with others. Local coffee shops, community places, listservs, mailing lists, social networks and forums are all great places to post details. Consider making a <a href="https://techkim.makes.org/thimble/maker-party-movie-poster">poster</a> with event information.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Sign Up Learners">
-                  <div>
-                    Your event can be drop-in or require registration but make sure there is an effective way for individuals to indicate they are attending and can get access to event information. Having a rough idea of how many learners will attend will give you a better idea of how to prepare stations, materials and mentors.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Schedule">
-                  <div>
-                    Plan your schedule ahead of time but be flexible; your participants might surprise you with interesting questions or challenges to explore. A checklist of things you will need for set-up/during/after is very useful. Don't forget to leave yourself adequate prep and clean-up time.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Press">
-                  <div>
-                    Media coverage is an essential part of promoting an event and showing off what you accomplish. Use the <a href="https://wiki.mozilla.org/Maker_Party/Communications/Press_Kit">Maker Party Press Kit</a> for guidelines on how to reach out to local media (newspapers, blogs, local news shows). Our <a href="https://wiki.mozilla.org/Maker_Party/Communications/Social_Media_Guide">social media toolkit</a> will also help you get more out of your posts online. Be sure to use hashtag <a href="https://twitter.com/search?f=realtime&q=%23makerparty">#MakerParty</a>. If your event is open to the public you may also want to invite community leaders and officials to attend.
-                  </div>
-                </EventDetail>
-              </div>
-            )
-          },
-          {
-            "head": "during event",
-            "content": (
-              <div>
-                <EventDetail head="Design a fun experience">
-                  <div>
-                    Your event should feel like a party! Create a welcoming, creative atmosphere that might include music, decorations and providing name tags. Starting your event with an offline activity or icebreaker such as a <a href="https://mozteach.makes.org/thimble/how-to-run-a-spectrogram-icebreaker">spectrogram</a> is a great way to get everyone socializing. Raffles and challenges are good ways to keep the crowd engaged throughout the day.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Set-up">
-                  <div>
-                    When preparing for event, be sure to arrive early to set-up the space using this sample <a href="https://stuff.webmaker.org/teach.mozilla.org/Event_Set_Up_Checklist.pdf">event checklist</a>. This will allow you plenty of time in case unexpected issues arise. Use posters or other clear signage to indicate different activity stations. Make sure washrooms and exits are clearly marked. Remind facilitators to welcome newcomers, and make sure the volunteers know what their responsibilities are and who to ask if they have questions.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Activities">
-                  <div>
-                    Find activity kits and tools online that encourage hands-on making, collaborating and participation from your attendees. Here are <Link to="activities">some of ours</Link> to get you started.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Documentation">
-                  <div>
-                    Time flies when you’re having fun! It's often impossible to capture or get links of all the awesome things being made. Appoint a volunteer to be the photographer and make sure that activities are well-documented and that photos/video can be shared online. Use the <a href="https://twitter.com/search?f=realtime&q=%23makerparty">#MakerParty</a> hashtag so we can see them too.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Share-outs">
-                  <div>
-                    Set aside time for attendees to see each others' activities and exchange ideas. At the end of the event, bring everyone back together in a circle. Celebrate what participants made and invite a few people to share their work with the entire group.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Wrap-up">
-                  <div>
-                    You’re done! Make sure to check the <a href="https://stuff.webmaker.org/teach.mozilla.org/Event_Take_Down_Checklist.pdf">event take down checklist</a> to make sure you’ve covered your bases. Do your best to make sure the space is in the same condition (if not better) than when you arrived.
-                  </div>
-                </EventDetail>
-              </div>
-            )
-          },
-          {
-            "head": "after event",
-            "content": (
-              <div>
-                <EventDetail head="Feedback">
-                  <div>
-                    Ask participants for feedback. What did they learn? What did they enjoy? If they have suggestions for improvement, be sure to take notes or invite them to blog about it.  Provide links to websites where they can continue learning new skills. Later, conduct a debrief with participating organizations and volunteers, so your next event can be even better.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Share what you learned">
-                  <div>
-                    Post pictures of your event on Twitter and Flickr using the hashtag <a href="https://twitter.com/search?f=realtime&q=%23makerparty">#MakerParty</a> or write a blog post using our <a href="https://stuff.webmaker.org/teach.mozilla.org/Sample_Blog_Post_Guidelines.pdf">sample guidelines</a> to share a deeper reflection.
-                  </div>
-                </EventDetail>
-                <EventDetail head="Thank you’s">
-                  <div>
-                    <a href="https://stuff.webmaker.org/teach.mozilla.org/Sample_Thank_You_Letters.pdf">Send a heartfelt</a> "Thank you!" to participating organizations, volunteers, facilitators, participants, partners, the venue and anyone else who helped make your event a success. Include photos, videos and links to the work that was created. Ask for feedback on the event and discuss next steps to getting involved, learning more or hosting a future event.
-                  </div>
-                </EventDetail>
-              </div>
-            )
-          }
-        ]}/>
-
-        <h2 id="event-support">Event Support</h2>
-        <p>Get help from the community and our staff.</p>
-
-        <div className="event-support-container">
-          <div className="event-support-panel">
-            <ImageTag
-            width={399} height={254}
-            src1x="/img/pages/event-resources/ask-the-community-image.png"
-            src2x="/img/pages/event-resources/ask-the-community-image@2x.png"
-            alt="ask the community image"/>
-            <h3 className="event-support-header">Ask The Community</h3>
-            <p>
-              We have a large, global community of people like you who have hosted events and have lots of advice to share. Visit our <a href="https://discourse.webmaker.org/c/events">discussion forum</a> to ask questions or share your own advice and experience with others.
-            </p>
-          </div>
-          <div className="event-support-panel">
-            <ImageTag
-            width={399} height={254}
-            src1x="/img/pages/event-resources/request-support-image.png"
-            src2x="/img/pages/event-resources/request-support-image@2x.png"
-            alt="request support image"/>
-            <h3 className="event-support-header">Request Support</h3>
-            <p>
-              Still can't find an answer to your question? Our team is here to help you with all things Maker Party. <a href={"mailto:"+config.TEACH_THE_WEB_EMAIL}>Contact us</a> and we will get back to you as soon as possible.
-            </p>
-          </div>
-        </div>
-
-      </div>
-    );
-  }
-});
-
-module.exports = EventsResources;
+// TODO: clean up indirection
+module.exports = require('./event-resources');
