+++ pages/event-resources/RemixLink.jsx
@@ -1,0 +1,11 @@
+var React = require('react');
+
+var RemixLink = React.createClass({
+  render: function() {
+    return (
+      <a className="remix-link" href={this.props.href}>Remix</a>
+    );
+  }
+});
+
+module.exports = RemixLink;
