+++ pages/event-resources/logoassets.jsx
@@ -1,0 +1,69 @@
+var React = require('react');
+var LogoAsset = require('./LogoAsset.jsx');
+var LogoAssetLink = require('./LogoAssetLink.jsx');
+var RemixLink = require('./RemixLink.jsx');
+
+module.exports = (
+  <div className="row">
+    <LogoAsset head="Maker Party Logo"
+    alt="Maker Party Logo Image"
+    src1x="/img/pages/event-resources/resource-thumbnails-01.png"
+    src2x="/img/pages/event-resources/resource-thumbnails-01@2x.png">
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyLogo.eps">Download EPS</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyLogo.png">Download PNG</LogoAssetLink>
+    </LogoAsset>
+    <LogoAsset head="Brand Palette"
+    alt="Brand Palette Image"
+    src1x="/img/pages/event-resources/resource-thumbnails-03.png"
+    src2x="/img/pages/event-resources/resource-thumbnails-03@2x.png">
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyBrandPalette.eps">Download EPS</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyBrandPalette.png">Download PNG</LogoAssetLink>
+    </LogoAsset>
+    <LogoAsset head="Poster"
+    alt="Poster Image"
+    src1x="/img/pages/event-resources/resource-thumbnails-07.png"
+    src2x="/img/pages/event-resources/resource-thumbnails-07@2x.png">
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyBanner.eps">Download EPS</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyBanner.png">Download PNG</LogoAssetLink>
+    </LogoAsset>
+    <LogoAsset head="Desktop Wallpaper"
+    alt="Desktop Wallpaper Image"
+    src1x="/img/pages/event-resources/resource-thumbnails-10.png"
+    src2x="/img/pages/event-resources/resource-thumbnails-10@2x.png">
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-320x480.jpg">320px x 480px</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-640x1136.jpg">640px x 1136px</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-768x1280.jpg">768px x 1280px</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-1920x1200.jpg">1920px x 1200px</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyWallpaper-1-2560x1440.jpg">2560px x 1440px</LogoAssetLink>
+    </LogoAsset>
+    <LogoAsset head="Certificate"
+    alt="Certificate Image"
+    src1x="/img/pages/event-resources/resource-thumbnails-08.png"
+    src2x="/img/pages/event-resources/resource-thumbnails-08@2x.png">
+      <RemixLink href="https://d157rqmxrxj6ey.cloudfront.net/mhasan/13029/ ">Download EPS</RemixLink>
+    </LogoAsset>
+    <LogoAsset head="Table Cloth Design"
+    alt="Table Cloth Design Image"
+    src1x="/img/pages/event-resources/resource-thumbnails-04.png"
+    src2x="/img/pages/event-resources/resource-thumbnails-04@2x.png">
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyTableClothDesign.eps">Download EPS</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyTableClothDesign.png">Download PNG</LogoAssetLink>
+    </LogoAsset>
+    <LogoAsset head="Buttons"
+    alt="Buttons Image"
+    src1x="/img/pages/event-resources/resource-thumbnails-06.png"
+    src2x="/img/pages/event-resources/resource-thumbnails-06@2x.png">
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyButtons.eps">Download EPS</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/MakerPartyButtons.png">Download PNG</LogoAssetLink>
+    </LogoAsset>
+    <LogoAsset head="KUMI Papertoy"
+    alt="KUMI Papertoy Image"
+    src1x="/img/pages/event-resources/resource-thumbnails-09.png"
+    src2x="/img/pages/event-resources/resource-thumbnails-09@2x.png">
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/KUMI-Papertoy-yellow.pdf">Yellow T-sirt</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/KUMI-Papertoy-blue.pdf">Teal T-shirt</LogoAssetLink>
+      <LogoAssetLink href="https://stuff.webmaker.org/teach.mozilla.org/KUMI-Papertoy-blank-tees.pdf">Design Your Own T-shirt</LogoAssetLink>
+    </LogoAsset>
+  </div>
+);
+
